Simplify dims/shape handling, remove component Inner type (#252)
* chore(nox): use shape to consolidate Dim traits

Make everything work with the same shape type.


* chore: consolidate on using i64 for shape


* chore: remove unused Inner assoc type of Component


---------

+++ libs/xla-rs/src/error.rs
@@ -12,7 +12,7 @@ pub enum Error {
     /// Incorrect number of elements.
     #[error("wrong element count {element_count} for dims {dims:?}")]
     WrongElementCount {
-        dims: Vec<usize>,
+        dims: Vec<i64>,
         element_count: usize,
     },
